# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Artemis Application Server API
  contact: {name: Stephan Krusche, email: krusche@tum.de}
  version: "10.0"
servers:
- {url: 'http://localhost:8080', description: Generated server url}
...
paths:
...
  /api/hyperion/authoring-runs/access:
    post:
      tags: [hyperion-authoring-run]
      operationId: checkAuthoringRunAccess
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items: {type: string, maxLength: 128, minLength: 1}
              maxItems: 500
              minItems: 1
        required: true
      responses:
        '200':
          description: Currently authorized run identities
          content:
            application/json:
              schema:
                type: array
                items: {type: string, maxLength: 128}
                maxItems: 500
...
  /api/hyperion/programming-exercises/{exerciseId}/generation/runs/{runId}:
    get:
      tags: [hyperion-authoring-run]
      operationId: getAuthoringRunStatus
      parameters:
      - name: exerciseId
        in: path
        required: true
        schema: {type: integer, format: int64}
      - name: runId
        in: path
        required: true
        schema: {type: string}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ExerciseGenerationStatus'}
...
  /api/hyperion/authoring-runs:
    get:
      tags: [hyperion-authoring-run]
      operationId: getAuthoringRuns
      parameters:
      - name: beforeId
        in: query
        required: false
        schema: {type: integer, format: int64, exclusiveMinimum: 0}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthoringRunPage'}
...
components:
  schemas:
...
    ExerciseGenerationFileChange:
      type: object
      description: A file changed by the generation agent
      properties:
        type:
          type: string
          description: Constant discriminator identifying a file change on the shared
            topic
          enum: [FILE_CHANGE]
        path: {type: string, description: Workspace-relative file path}
        repo:
          type: string
          description: Owning repository bucket
          enum: [solution, template, tests, other]
        action:
          type: string
          description: The successful file tool operation
          enum: [write, edit, delete]
        turn: {type: integer, format: int32, description: The agent turn the change
            happened on}
        timestamp: {type: string, format: date-time, description: The moment the change
            was produced}
      required: [action, path, repo, timestamp, turn, type]
...
    ExerciseGenerationUsage:
      type: object
      properties:
        modelCalls: {type: integer, format: int64}
        toolCalls: {type: integer, format: int64}
        agentTurns: {type: integer, format: int64}
        attempts: {type: integer, format: int64}
        inputTokens: {type: integer, format: int64}
        outputTokens: {type: integer, format: int64}
        cachedInputTokens: {type: integer, format: int64}
        cachedInputTokensComplete: {type: boolean}
        estimatedCostEur: {type: number, format: double}
        estimatedCostEurComplete: {type: boolean}
        models:
          type: array
          items: {type: string}
        providerRequestIds:
          type: array
          items: {type: string}
        providerRequestIdsComplete: {type: boolean}
      required: [agentTurns, attempts, cachedInputTokens, cachedInputTokensComplete,
        estimatedCostEur, estimatedCostEurComplete, inputTokens, modelCalls, models,
        outputTokens, providerRequestIds, providerRequestIdsComplete, toolCalls]
...
    ExerciseGenerationEvent:
      type: object
      description: A progress event streamed to the instructor while an agentic whole-exercise
        generation or adaptation runs
      properties:
        type:
          type: string
          description: The event kind
          enum: [STARTED, PROGRESS, DONE, CANCELLED, ERROR]
        message: {type: string, description: Human-readable progress or result message}
        phase:
          type: string
          description: Stable instructor-facing phase of the generation journey
          enum: [PREPARING, DESIGNING, VERIFYING, REVIEWING, REPAIRING, SAVING]
        completionStatus:
          type: string
          description: On a terminal DONE event, whether the run succeeded, needs
            review, or partially completed
          enum: [SUCCESS, NEEDS_REVIEW, PARTIAL]
        verdict: {$ref: '#/components/schemas/ExerciseGenerationVerdict', description: 'On
            a terminal event, the structured verification verdict'}
        liveExerciseChanged: {type: boolean, description: 'On a terminal event, whether
            the live exercise changed and open editors should refresh'}
        savedRepositoryCommits:
          type: object
          additionalProperties: {type: string}
          description: Exact saved commit hashes keyed by repository name
        savedExerciseVersionId: {type: integer, format: int64, description: 'The exact
            saved exercise version id, on a successful terminal event'}
        terminationReason:
          type: string
          description: On a terminal event, why the generation run ended
          enum: [CONVERGED, REPAIR_BUDGET_EXHAUSTED, ATTEMPT_CAP_REACHED, MECHANICAL_REPAIR_EXHAUSTED,
            POST_REPAIR_CORRECTION_EXHAUSTED, NO_SCHEDULABLE_SURFACE, REVIEW_UNAVAILABLE,
            UNCHANGED_CANDIDATE_RESUBMITTED, REPAIR_DID_NOT_IMPROVE, NO_ADMISSIBLE_CONCEPT,
            CONCEPT_ADMITTED_WITH_FINDINGS, CANCELLED, DEADLINE_EXCEEDED, TOKEN_BUDGET_EXHAUSTED,
            AGENT_ERROR, ENVIRONMENT_UNAVAILABLE, RUN_FAILED, NOT_STARTED]
        repairRound: {$ref: '#/components/schemas/ExerciseGenerationRepairRound',
          description: 'On a repair-round progress event, that round''s finding bookkeeping'}
        activity: {$ref: '#/components/schemas/ExerciseGenerationActivity', description: 'On
            a progress event produced from inside the agent loop, what the run was
            doing at that moment'}
        liveUsage: {$ref: '#/components/schemas/ExerciseGenerationLiveUsage', description: 'On
            an activity or phase event, what the run has spent up to that moment'}
        timestamp: {type: string, format: date-time, description: The moment the event
            was produced}
      required: [timestamp, type]
...
    ExerciseGenerationStatus:
      type: object
      properties:
        jobId: {type: string}
        running: {type: boolean}
        mode:
          type: string
          enum: [GENERATE, ADAPT]
        events:
          type: array
          items: {$ref: '#/components/schemas/ExerciseGenerationEvent'}
        fileChanges:
          type: array
          items: {$ref: '#/components/schemas/ExerciseGenerationFileChange'}
        revertAvailable: {type: boolean}
        revertJobId: {type: string}
        revertMode:
          type: string
          enum: [GENERATE, ADAPT]
        ownedByCaller: {type: boolean}
        cancellable: {type: boolean}
        specDocument: {type: string}
        usage: {$ref: '#/components/schemas/ExerciseGenerationUsage'}
        accountingState:
          type: string
          description: Whether the reported usage is a complete account of a generation
            run's provider spend
          enum: [PENDING, COMPLETE, INCOMPLETE]
        effortProfile: {type: string}
        artifactsRetained: {type: boolean, description: Whether a current or retained
            candidate snapshot from this run is readable}
        input: {$ref: '#/components/schemas/ExerciseGenerationInput'}
        run: {$ref: '#/components/schemas/AuthoringRun'}
      required: [accountingState, artifactsRetained, cancellable, events, fileChanges,
        jobId, ownedByCaller, revertAvailable, running]
...
    ExerciseGenerationActivity:
      type: object
      description: What a generation run is doing right now, so a long synchronous
        model call is visible instead of silent
      properties:
        step: {type: string, description: 'Machine-readable substep the run is inside
            (concept, spec, artifacts, statement), or null when no substep applies'}
        attempt: {type: integer, format: int32, description: The 1-based authoring
            attempt}
        turn: {type: integer, format: int32, description: The 1-based agent turn inside
            the current agent session}
        waitingOnModel: {type: boolean, description: Whether this event was emitted
            immediately before a provider call the run is now waiting on}
        modelCalls: {type: integer, format: int32, description: How many provider
            calls the run has completed so far}
        toolCalls: {type: integer, format: int32, description: How many tool calls
            the run has executed so far}
        filesWritten: {type: integer, format: int32, description: 'How many successful
            file writes, edits, and deletions the run has made so far'}
      required: [attempt, filesWritten, modelCalls, toolCalls, turn, waitingOnModel]
...
    ExerciseGenerationVerdict:
      type: object
      description: The structured outcome of the differential verification, so the
        client can show which gates passed without parsing prose
      properties:
        mechanicallyVerified: {type: boolean, description: Whether the exercise passed
            mechanical verification}
        solutionPassed: {type: boolean, description: Whether the solution passed all
            its tests}
        templateFailed: {type: boolean, description: Whether the template compiled
            but correctly failed the tests}
        testCount: {type: integer, format: int32, description: The number of tests
            discovered}
        reasons:
          type: array
          description: Human-readable explanations of any failed gate (empty when
            mechanically verified)
          items: {type: string}
      required: [mechanicallyVerified, reasons, solutionPassed, templateFailed, testCount]
...
    ExerciseGenerationFeedback:
      type: object
      properties:
        targetType: {type: string}
        filePath: {type: string}
        lineNumber: {type: integer, format: int32}
        comments:
          type: array
          items: {type: string}
...
    AuthoringRunPage:
      type: object
      properties:
        runs:
          type: array
          items: {$ref: '#/components/schemas/AuthoringRun'}
        nextBeforeId: {type: integer, format: int64}
      required: [runs]
...
    ExerciseGenerationLiveUsage:
      type: object
      description: What a generation run has spent so far, streamed while it runs
      properties:
        inputTokens: {type: integer, format: int64, description: Prompt tokens billed
            so far}
        outputTokens: {type: integer, format: int64, description: Completion tokens
            billed so far}
        cachedInputTokens: {type: integer, format: int64, description: The share of
            the prompt tokens the provider served from its cache}
        billableTokens: {type: integer, format: int64, description: 'Tokens charged
            against the run''s token budget, with cached input discounted'}
        tokenBudget: {type: integer, format: int64, description: The token ceiling
            this run was admitted against}
        modelCalls: {type: integer, format: int64, description: Provider responses
            accounted for so far}
        estimatedCostEur: {type: number, format: double, description: 'Estimated cost
            in EUR, absent when any model used has no configured price'}
        estimatedCostComplete: {type: boolean, description: Whether a price was configured
            for every response counted so far}
      required: [billableTokens, cachedInputTokens, estimatedCostComplete, inputTokens,
        modelCalls, outputTokens, tokenBudget]
...
    ExerciseGenerationInput:
      type: object
      properties:
        prompt: {type: string}
        reviewFeedback:
          type: array
          items: {$ref: '#/components/schemas/ExerciseGenerationFeedback'}
        sourceExerciseId: {type: integer, format: int64}
...
    ExerciseGenerationRepairRound:
      type: object
      description: Per-round finding bookkeeping of the semantic repair loop, so a
        run's transcript shows whether repairing drains findings
      properties:
        round: {type: integer, format: int32, description: The 1-based index of this
            review round within the run}
        attempt: {type: integer, format: int32, description: The 1-based authoring
            attempt whose candidate was reviewed}
        blocking: {type: integer, format: int32, description: How many of this round's
            findings block acceptance}
        advisory: {type: integer, format: int32, description: How many of this round's
            findings are advisory}
        carriedOver: {type: integer, format: int32, description: How many of this
            round's findings were already present in the previous round}
        drained: {type: integer, format: int32, description: How many of the previous
            round's findings are gone in this round}
        fresh: {type: integer, format: int32, description: How many of this round's
            findings are new since the previous round}
      required: [advisory, attempt, blocking, carriedOver, drained, fresh, round]
...
    AuthoringRun:
      type: object
      properties:
        jobId: {type: string}
        exerciseId: {type: integer, format: int64}
        sourceExerciseId: {type: integer, format: int64}
        courseId: {type: integer, format: int64}
        exerciseTitle: {type: string}
        kind:
          type: string
          enum: [CREATE, ADAPT, VARIANT]
        status:
          type: string
          enum: [QUEUED, SAVED, NEEDS_REVIEW, PARTIAL, ERROR, CANCELLED, UNKNOWN]
        running: {type: boolean}
        startedAt: {type: string, format: date-time}
        finishedAt: {type: string, format: date-time}
        beforeVersionId: {type: integer, format: int64}
        afterVersionId: {type: integer, format: int64}
        revertedAt: {type: string, format: date-time}
      required: [courseId, exerciseId, jobId, kind, running, startedAt, status]
...