# src/main/webapp/app/openapi/model/authoring-run-page.ts
/**
 * Artemis Application Server API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API Version: 10.0
 * Contact: krusche@tum.de
 *
 * NOTE: This file is auto-generated. Do not edit manually.
 */

import type { AuthoringRun } from './authoring-run';

export interface AuthoringRunPage {
    runs: Array<AuthoringRun>;
    nextBeforeId?: number;
}